requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
configparser==5.3.0
exceptiongroup==1.2.0
Faker==18.13.0
idna==3.7
importlib-metadata==6.7.0
iniconfig==2.0.0
jsonpath==0.82.2
packaging==24.0
pluggy==1.2.0
PyMySQL==1.1.0
pytest==7.4.4
pytest-ordering==0.6
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
